requirements.txt
python==3.12.11
fastapi==0.115.13
uvicorn==0.34.3
gunicorn==23.0.0
Pillow==11.2.1
requests==2.32.4
httpx==0.28.1
aio-pika==9.5.5
celery==5.5.3
redis==6.2.0
aiomysql==0.2.0
asyncpg==0.30.0
# psycopg2==2.9.10
sqlalchemy==2.0.41
python-dotenv==1.1.1
pydantic_settings==2.10.1
passlib==1.7.4
python-jose==3.5.0